Implement and test cost and usage command parsing

+++ src/main/scala/com/exercise/tariff/cli/Commands.scala
@@ -1,5 +1,5 @@
 package com.exercise.tariff.cli
 
 sealed trait Command
-case class CostCommand(power: Double, gas: Double) extends Command
-case class UsageCommand(tariffName: String, fuel: Fuel, targetMonthlySpend: Double) extends Command
+case class CostCommand(power: Int, gas: Int) extends Command
+case class UsageCommand(tariffName: String, fuel: Fuel, targetMonthlySpend: Int) extends Command
